requirements.txt
Django==1.5.8
PyYAML==3.11
South==1.0
argparse==1.2.1
django-mongodb-engine==0.5.2
djangorestframework==2.4.3
djangotoolbox==1.6.2
helper==2.4.1
newrelic-plugin-agent==1.3.0
pymongo==2.7.2
requests==2.4.3
simplejson==3.6.5
wsgiref==0.1.2
